Fix property test prop_aeson_canonical
Canonical JSON is not a proper subset of RFC 7159.

So for the property 'prop_aeson_canonical', where we check that
everything produced as canoncal JSON can be parsed by Aeson (which we
assume correctly implements RFC 7159), we have to tweak things to keep
us within the common subset of canoncal JSON and RFC 7159.
Specifically, canoncal JSON only escapes \ and ", and does not escape
any other non-printable characters.

So the tweak is to just omit non-printable characters from all strings
in this test. Thus we test only within the common subset.